feat(providers): attribute resolved indexer + consolidate provider & indexer health on dashboard
Indexer attribution:
- Thread an already-known indexer through AddToQueue so it is attributed at
  insert time instead of falling back to "Unknown". NZBLNK imports relay the
  resolver's indexer, and Stremio plays relay the Prowlarr indexer via the play
  URL — both previously dropped that information.

Dashboard / provider & indexer health:
- Provider Status table shows the assigned nickname (falling back to the host
  brand name) and a Data Usage column. The card keeps the original
  Network-icon "NNTP Providers" header, and the data-usage and speed-history
  graphs are merged into the same card as a section below the table (side by
  side, breakdown donuts removed, single divider line).
- Indexer Health is consolidated into a single card — header, tracked-indexer
  stat cards, filters, and the table all merged into one table.
- Move the provider card (table + graphs) and the Indexer Health card onto the
  Dashboard.
- Remove the Provider Health and Indexer Health tabs from the Health page
  (now File Health only).
- Stop dashboard icons from animating (provider speed, health activity, graph
  header, and indexer HUD icons are now static).
