+++ frontend/src/components/system/IndexerHealth/PruneStatsModal.tsx
