+++ internal/api/nzb_stremio_handlers.go
@@ -266,7 +266,7 @@ func (s *Server) handleNzbStreams(c *fiber.Ctx) error {
 	}
 
 	priority := database.QueuePriorityHigh
-	item, err := s.importerService.AddToQueue(ctx, tempPath, basePath, &category, &priority, nil, nil)
+	item, err := s.importerService.AddToQueue(ctx, tempPath, basePath, &category, &priority, nil, nil, nil)
 	if err != nil {
 		os.Remove(tempPath)
 		return RespondInternalError(c, "Failed to add NZB to queue", err.Error())
