+++ internal/importer/scanner/watcher.go
@@ -16,7 +16,7 @@ import (
 
 // WatchQueueAdder interface for adding items to the import queue from directory watcher
 type WatchQueueAdder interface {
-	AddToQueue(ctx context.Context, filePath string, relativePath *string, category *string, priority *database.QueuePriority, metadata *string, downloadID *string) (*database.ImportQueueItem, error)
+	AddToQueue(ctx context.Context, filePath string, relativePath *string, category *string, priority *database.QueuePriority, metadata *string, downloadID *string, indexer *string) (*database.ImportQueueItem, error)
 	IsFileInQueue(ctx context.Context, filePath string) (bool, error)
 }
 
@@ -290,7 +290,7 @@ func (w *Watcher) processNzb(ctx context.Context, watchRoot, filePath string) er
 
 	// Add to queue
 	priority := database.QueuePriorityNormal
-	item, err := w.queueAdder.AddToQueue(ctx, filePath, relativePath, category, &priority, nil, nil)
+	item, err := w.queueAdder.AddToQueue(ctx, filePath, relativePath, category, &priority, nil, nil, nil)
 	if err != nil {
 		return fmt.Errorf("failed to add to queue: %w", err)
 	}
